Moving optimizer base classes to the proper file.
