+++ testing_neuraxle/metaopt/test_automl.py
@@ -12,9 +12,9 @@ from neuraxle.metaopt.callbacks import (BestModelCheckpoint,
 from neuraxle.metaopt.context import AutoMLContext
 from neuraxle.metaopt.data.aggregates import Trial
 from neuraxle.metaopt.data.reporting import RoundReport, TrialReport
-from neuraxle.metaopt.validation import (GridExplorationSampler,
-                                         KFoldCrossValidationSplitter,
-                                         RandomSearchSampler,
+from neuraxle.metaopt.optimizer import (GridExplorationSampler,
+                                        RandomSearchSampler)
+from neuraxle.metaopt.validation import (KFoldCrossValidationSplitter,
                                          ValidationSplitter)
 from neuraxle.pipeline import Pipeline
 from neuraxle.steps.flow import ChooseStepElseIdentity
