+++ testing_neuraxle/test_context_logger.py
@@ -16,11 +16,12 @@ from neuraxle.logging.logging import (NEURAXLE_LOGGER_NAME, NeuraxleLogger,
 from neuraxle.metaopt.auto_ml import AutoML
 from neuraxle.metaopt.callbacks import ScoringCallback
 from neuraxle.metaopt.context import AutoMLContext
-from neuraxle.metaopt.repositories.json import HyperparamsOnDiskRepository
 from neuraxle.metaopt.data.vanilla import (DEFAULT_CLIENT, DEFAULT_PROJECT,
                                            ProjectDataclass, ScopedLocation,
                                            TrialSplitDataclass)
-from neuraxle.metaopt.validation import RandomSearchSampler, ValidationSplitter
+from neuraxle.metaopt.optimizer import RandomSearchSampler
+from neuraxle.metaopt.repositories.json import HyperparamsOnDiskRepository
+from neuraxle.metaopt.validation import ValidationSplitter
 from neuraxle.pipeline import Pipeline
 from neuraxle.steps.misc import FitTransformCounterLoggingStep
 from neuraxle.steps.numpy import MultiplyByN, NumpyReshape
